Replace Dubeolsik system with Unicode translation
The previous system output English characters, which would then be
converted to Korean letters and syllables through the computer's Korean
keyboard layout. This required the user to switch over to the Korean
layout first, before using this plugin.

The new system uses some arithmetic, and numbers assigned to all the
letters, to determine which of the ~11,000 Hangul syllable block code
points to return. See the full table of Hangul syllables here:

https://en.wikipedia.org/wiki/Hangul_Syllables

Here's how the algorithm works...

19 initial jamo:

0. ᄀ 4. ㄸ  8. ㅃ 12. ㅈ 16. ㅌ
1. ㄲ 5. ㄹ  9. ㅅ 13. ㅉ 17. ㅍ
2. ㄴ 6. ㅁ 10. ㅆ 14. ㅊ 18. ㅎ
3. ㄷ 7. ㅂ 11. ㅇ 15. ㅋ

21 medial jamo:

0. ㅏ  7. ㅖ 14. ㅝ
1. ㅐ  8. ㅗ 15. ㅞ
2. ㅑ  9. ㅘ 16. ㅟ
3. ㅒ 10. ㅙ 17. ㅠ
4. ㅓ 11. ㅚ 18. ㅡ
5. ㅔ 12. ㅛ 19. ㅢ
6. ㅕ 13. ㅜ 20. ㅣ

27 final jamo (28 including "None"):

0.     7. ㄷ 14. ㄿ 21. ㅇ
1. ㄱ  8. ㄹ 15. ㅀ 22. ㅈ
2. ㄲ  9. ㄺ 16. ㅁ 23. ㅊ
3. ㄳ 10. ㄻ 17. ㅂ 24. ㅋ
4. ㄴ 11. ㄼ 18. ㅄ 25. ㅌ
5. ㄵ 12. ㄽ 19. ㅅ 26. ㅍ
6. ㄶ 13. ㄾ 20. ㅆ 27. ㅎ

We plug those numbers into this formula:

44032 + initial * 588 + medial * 28 + final

The first number gets us to the start of the Hangul set of characters in
the Unicode system, and the rest offset us from there to the syllable
that matches the parts. The 28 final characters repeat over and over,
for each medial, and the medials repeat every 21 medials * 28 finals, or
every 588 characters, and each of those repeats starts with a different
initial character.

Currently, I don't now how to slot in the "tense" final consonants—I
just don't understand how those work in Hangul, so their contributions
are currently set to 0—and I can't see how the compound final consonants
are meant to be created. For the latter, I watched the demo video,
explored the code, and did some looking around online, and never really
saw them in use anywhere. When I understand these things, I can update
the code.