Fix some errors in vowel entries
A TH variant chord for ㄸ was accidentally in the regular consonants
dictionary. The O chord for ㅗ was in the regular and y-vowel
dictionaries, but the OE variant was only in the regular, where it
denotes the ㅛ vowel. There was an incorrect AOU entry for ㅠ, which
should only work for ㅡ. Finally, the entry for "EU"=ㅣwas missing.