Remove duplicate vowel
AOU was being used for ㅝ and ㅡ, and since ㅡ already has a chord, AU,
I removed the AOU → ㅡ mapping. In the process, I also realized that
we're not currently using AOE for anything. We're using 14 of the 15
vowel chords (not including the 16th chord, which is no keys pressed).
This means we could give that chord to one of the two remaining vowels
that doesn't have a chord. We could also give the lack of any chord keys
to the final chord, and then have them all. We wouldn't be able to write
that chord on its own, but that could be filled in with a custom
sequence, or we could just give it to the least common one.

I also removed the empty case for starting vowels, which I think is
right, but will need some more testing to confirm.