Fix lone ending consonant usage
Whoops, typo. Now we have ending consonants that can be used on their
own. Stroke the starting TKPW chord, and it outputs ㄱ. Stroke the AOU
chord, and you get ㅝ. Stroke the ending -FR chord, you get ㅀ.

If you want a vowel that can have Y added, like ㅏ → ㅑ, just add the
star key, so A is ㅏ, and A* is ㅑ. This is requirement is due to the
fact that adding the Y sound is normally done as part of the initial
consonant, as in KA → 가, vs KWA → 갸, so we can't write those vowels on
their own otherwise.