Add combined version of special, tense systems
The way things are currently laid out, the easiest way to get this last
piece of the puzzle in place was to create a tense+special dictionary,
which combines the chords of the special dict with the doubled letters
of the tense dict. This seems to work now, so using a special chord, say
SH instead of S—to get the version of an initial consonant with a
y-vowel after it—along with the star—to get the tense (doubled) version
of that consonant—will finally output both things together in the
syllable. That means all the things work. There's just a little bit of
cleanup and vowel decisions remaining.